Move most of pem handling to its own file
